pyproject.toml
[project]
name = "natural-language-processing"
version = "0.1.0"
description = "Natural Language Processing."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black[jupyter]>=26.1.0",
    "gensim>=4.4.0",
    "matplotlib>=3.10.8",
    "polars>=1.38.1",
    "pre-commit>=4.5.1",
    "pyyaml>=6.0.3",
    "rich>=14.3.2",
    "scikit-learn>=1.8.0",
    "torch>=2.10.0",
    "typing>=3.10.0.0",
]


[dependency-groups]
lint = [
    "black",
    "black[jupyter]"
]

notebooks = [
    "ipykernel>=7.2.0",
    "ipywidgets>=8.1.8",
]

[tool.black]
line-length = 79
target-version = ['py311']
include = '\.pyi?$'
